requirements.txt
fastapi
uvicorn[standard]
pydantic
strawberry-graphql[fastapi]